pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plagentic"
version = "0.1.0"
description = "Agentic AI Infrastructure Provisioning Platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "GNUv3"}
authors = [
    {name = "smadi0x86", email = "smadi0x86dev@gmail.com"}
]
keywords = ["ai", "agents", "infrastructure", "provisioning", "cloud", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",

    "Programming Language :: Python :: 3.11",   
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "requests>=2.32.0",
    "urllib3>=1.26.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "browser-use==0.1.40",
    "playwright>=1.54.0",
]

[project.optional-dependencies]
full = [
    "browser-use>=0.1.40",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
# CLI Entry Point - Command Line Interface
plagentic = "plagentic.cli.cli:app"

[project.urls]
Homepage = "https://github.com/smadi0x86/plagentic"
Documentation = ""
Repository = "https://github.com/smadi0x86/plagentic"
Issues = "https://github.com/smadi0x86/plagentic/issues"

[tool.hatch.build.targets.wheel]
packages = ["plagentic"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "browser_use.*",
    "yaml.*",
]
ignore_missing_imports = true
